+++ db/seeds/group.rb
@@ -78,8 +78,14 @@ group_attrs = [
 ]
 
 group_attrs.each do |attr_title|
-    AssociatedAttr.find_or_create_by(attr_title: attr_title) do |attr|
-      # Set other attributes if necessary, for example:
-
+    associated_attr = AssociatedAttr.find_or_initialize_by(attr_title: attr_title)
+  
+    # Get existing endpoints or initialize an empty array
+    existing_endpoints = associated_attr.associated_endpoints || []
+  
+    # Only append 'ActionLog' if it's not already present
+    unless existing_endpoints.include?("Group")
+      new_endpoints = existing_endpoints + ["Group"] # Append only if it doesn't exist
+      associated_attr.update(associated_endpoints: new_endpoints)
     end
   end